refactor(#1749): move Python refresh() into refresh.rs to clear file-size ratchet
database.rs grew 946 -> 997 when Database.refresh() was added; the campsite
ratchet FAILs growth of an over-threshold file. Split by responsibility per
epic #1116: move the refresh() #[pymethods] block into refresh.rs alongside
RefreshReport (mirrors bindings/node/src/refresh.rs). Enable pyo3
multiple-pymethods so the second impl block coexists with database.rs; the
inventory dependency it pulls in builds fine on this toolchain. Public API is
unchanged (db.refresh() works, RefreshReport still importable from cqlite).

database.rs now 946 lines (== merge-base), ratchet PASS.
